requirements.txt
asgiref==3.2.3
astroid==2.3.3
attrs==19.3.0
Automat==20.2.0
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==8.1.3
colorama==0.4.3
constantly==15.1.0
cryptography==2.9.2
cssselect==1.1.0
decorator==5.1.1
Django==3.0.2
django-crispy-forms==1.9.2
django-filter==2.2.0
future==0.18.2
geocoder==1.38.1
get==0.0.0
hyperlink==19.0.0
idna==2.10
importlib-metadata==4.11.4
incremental==17.5.0
isodate==0.6.0
isort==4.3.21
itemadapter==0.1.0
lazy-object-proxy==1.4.3
lxml==4.5.1
mccabe==0.6.1
parsel==1.6.0
Pillow==7.0.0
post==0.0.0
Protego==0.1.16
psycopg2==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
PyHamcrest==2.0.2
pylint==2.4.4
pylint-django==2.0.15
pylint-plugin-utils==0.6
pyOpenSSL==19.1.0
pytz==2019.3
query-string==0.0.0
queuelib==1.5.0
ratelim==0.1.6
requests==2.24.0
Scrapy==2.2.0
service-identity==18.1.0
six==1.14.0
sqlparse==0.3.0
Twisted==20.3.0
typed-ast==1.4.1
typing-extensions==4.2.0
urllib3==1.25.9
w3lib==1.22.0
wrapt==1.11.2
zipp==3.8.0
zope.interface==5.1.0
